+++ pkg/helpers/README.md
@@ -1,0 +1,3 @@
+# What does this folder contain?
+
+This folder contains code performing several smaller tasks.
