+++ pkg/output/README.md
@@ -1,0 +1,3 @@
+# What does this folder contain?
+
+This folder contains code used to print to terminal.
